requirements.txt
asgiref==3.8.1
black==25.1.0
boto3==1.38.35
botocore==1.38.35
click==8.1.8
colorama==0.4.6
Django==5.2
django-rest-framework==0.1.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
jmespath==1.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
